Refactor directories in authn-jwt
