--- app/domain/authentication/authn_jwt/create_signing_key_interface.rb
@@ -1,68 +1,0 @@
-module Authentication
-  module AuthnJwt
-    # Factory that returns the interface implementation of FetchSigningKey
-    CreateSigningKeyInterface ||= CommandClass.new(
-      dependencies: {
-        fetch_provider_uri: Authentication::AuthnJwt::FetchProviderUriSigningKey,
-        fetch_jwks_uri: Authentication::AuthnJwt::FetchJwksUriSigningKey,
-        logger: Rails.logger
-      },
-      inputs: %i[authentication_parameters]
-    ) do
-
-      def call
-        create
-      end
-
-      private
-
-      def create
-        @logger.debug(LogMessages::Authentication::AuthnJwt::ValidatingJwtSigningKeyConfiguration.new)
-        validate_key_configuration
-
-        if provider_uri_has_valid_configuration?
-          @logger.debug(LogMessages::Authentication::AuthnJwt::FetchingProviderUriSigningKey.new)
-          fetch_provider_uri_signing_key
-        elsif jwks_uri_has_valid_configuration?
-          @logger.debug(LogMessages::Authentication::AuthnJwt::FetchingJwksUriSigningKey.new)
-          fetch_jwks_uri_signing_key
-        end
-      end
-
-      def validate_key_configuration
-        if (provider_uri_has_valid_configuration? and jwks_uri_has_valid_configuration?) or
-          (!provider_uri_has_valid_configuration? and !jwks_uri_has_valid_configuration?)
-          raise Errors::Authentication::AuthnJwt::InvalidUriConfiguration.new(
-            PROVIDER_URI_RESOURCE_NAME,
-            JWKS_URI_RESOURCE_NAME
-          )
-        end
-      end
-
-      def provider_uri_has_valid_configuration?
-        @provider_uri_has_valid_configuration ||= fetch_provider_uri_signing_key.has_valid_configuration?
-      end
-
-      def jwks_uri_has_valid_configuration?
-        @jwks_uri_has_valid_configuration ||= fetch_jwks_uri_signing_key.has_valid_configuration?
-      end
-
-      def fetch_provider_uri_signing_key
-        @fetch_provider_uri_signing_key ||= @fetch_provider_uri.new(authentication_parameters: @authentication_parameters,
-                                                                    logger: Rails.logger,
-                                                                    fetch_required_secrets: Conjur::FetchRequiredSecrets.new,
-                                                                    resource_class: ::Resource,
-                                                                    discover_identity_provider: Authentication::OAuth::DiscoverIdentityProvider.new)
-      end
-
-      def fetch_jwks_uri_signing_key
-        @fetch_jwks_uri_signing_key ||= @fetch_jwks_uri.new(authentication_parameters: @authentication_parameters,
-                                                            logger: Rails.logger,
-                                                            fetch_required_secrets: Conjur::FetchRequiredSecrets.new,
-                                                            resource_class: ::Resource,
-                                                            http_lib: Net::HTTP,
-                                                            create_jwks_from_http_response: CreateJwksFromHttpResponse.new)
-      end
-    end
-  end
-end
