--- app/domain/authentication/authn_jwt/fetch_signing_key_interface.rb
@@ -1,8 +1,0 @@
-module Authentication
-  module AuthnJwt
-    class FetchSigningKeyInterface
-      def create; end
-      def has_valid_configuration?; end
-    end
-  end
-end
