--- app/domain/authentication/authn_jwt/identity_provider_interface.rb
@@ -1,10 +1,0 @@
-module Authentication
-  module AuthnJwt
-    class IdentityProviderInterface
-      def initialize(authentication_parameters); end
-      def provide_jwt_identity; end
-      def identity_available?; end
-    end
-  end
-end
-
