+++ app/domain/authentication/authn_jwt/identity_providers/identity_from_url_provider.rb
@@ -1,0 +1,21 @@
+module Authentication
+  module AuthnJwt
+    module IdentityProviders
+      # Provides jwt identity from information in the URL
+      class IdentityFromUrlProvider < IdentityProviderInterface
+        def initialize(authentication_parameters)
+          @authentication_parameters = authentication_parameters
+        end
+
+        def provide_jwt_identity
+          raise Errors::Authentication::AuthnJwt::NoRelevantIdentityProvider.new unless self.identity_available?
+          @authentication_parameters.username
+        end
+
+        def identity_available?
+          !@authentication_parameters.username.blank?
+        end
+      end
+    end
+  end
+end
