+++ app/domain/authentication/authn_jwt/orchestrate_jwt_authentication.rb
@@ -6,8 +6,8 @@ module Authentication
 
     OrchestrateAuthentication ||= CommandClass.new(
       dependencies: {
-        validate_uri_based_parameters: Authentication::AuthnJwt::ValidateUriBasedParameters.new,
-        jwt_configuration_factory: ConfigurationFactory.new,
+        validate_uri_based_parameters: Authentication::AuthnJwt::InputValidation::ValidateUriBasedParameters.new,
+        jwt_configuration_factory: Authentication::AuthnJwt::VendorConfigurations::ConfigurationFactory.new,
         jwt_authenticator: Authentication::AuthnJwt::Authenticator.new,
         logger: Rails.logger
       },
