+++ app/domain/authentication/authn_jwt/signing_key/fetch_signing_key_interface.rb
@@ -1,0 +1,10 @@
+module Authentication
+  module AuthnJwt
+    module SigningKey
+      class FetchSigningKeyInterface
+        def create; end
+        def has_valid_configuration?; end
+      end
+    end
+  end
+end
